+++ CHANGELOG.md
@@ -8,6 +8,8 @@
     * Handle default `batch_size` value for local runner (\#2949).
     * Rename `SudoSlurmRunner` into `SlurmSudoRunner` (\#2980).
     * Ignore some SLURM/stderr patterns when populating `executor_error_log` (\#2984).
+* App Settings:
+    * Transform `DB_ECHO` configuration variable from boolean to `Literal['true', 'false']` (\#2985).
 
 # 2.17.1